Organize glyph package
